refactor: consolidate and enhance VersionedIBucketTester for improved version handling
