+++ python/tests/bucket_tester.py
@@ -17,7 +17,6 @@ import pyarrow.parquet as pq
 from streamerate import slist
 from streamerate import stream as sstream
 
-from bucketbase import VersionedMinioBucket
 from bucketbase.ibucket import AsyncObjectWriter, IBucket
 
 
@@ -78,78 +77,6 @@ class FailingStream(io.IOBase):
 
     def writable(self) -> bool:
         return True
-
-
-class VersionedIBucketTester:
-    def __init__(self, storage: VersionedMinioBucket, test_case: TestCase) -> None:
-        self.storage = storage
-        self.test_case = test_case
-        self.us = uuid.uuid4().hex
-        self._tracked_names: list[PurePosixPath] = []
-
-    def cleanup(self) -> None:
-        for name in self._tracked_names:
-            self.storage.remove_object_with_versions(name)
-
-    def _track(self, name: PurePosixPath) -> PurePosixPath:
-        self._tracked_names.append(name)
-        return name
-
-    def test_full_cycle_object_versions_after_overwrite(self) -> None:
-        path = self._track(PurePosixPath(f"dir{self.us}/versioned.txt"))
-
-        self.storage.put_object(path, b"old content")
-        self.storage.put_object(path, b"new content")
-
-        versions = self.storage.list_object_versions(path)
-        object_versions = [version for version in versions if not version.is_delete_marker]
-        latest_versions = [version for version in object_versions if version.is_latest]
-        old_versions = [version for version in object_versions if not version.is_latest]
-
-        self.test_case.assertIsInstance(versions, slist)
-        self.test_case.assertEqual(2, len(object_versions))
-        self.test_case.assertEqual(1, len(latest_versions))
-        self.test_case.assertEqual(1, len(old_versions))
-        self.test_case.assertEqual(b"new content", self.storage.get_object(path))
-        self.test_case.assertEqual(b"old content", self.storage.get_object_version(path, old_versions[0].version_id))
-        with self.storage.get_object_version_stream(path, old_versions[0].version_id) as stream:
-            self.test_case.assertEqual(b"old content", stream.read())
-
-        errors = self.storage.remove_objects([path])
-        versions_after_delete = self.storage.list_object_versions(path)
-        delete_markers = [version for version in versions_after_delete if version.is_delete_marker]
-
-        self.test_case.assertEqual([], list(errors))
-        self.test_case.assertFalse(self.storage.exists(path))
-        self.test_case.assertEqual(1, len(delete_markers))
-        self.test_case.assertTrue(delete_markers[0].is_latest)
-        self.test_case.assertEqual(b"old content", self.storage.get_object_version(path, old_versions[0].version_id))
-        with self.test_case.assertRaises(FileNotFoundError):
-            self.storage.get_object(path)
-        with self.test_case.assertRaises(FileNotFoundError):
-            self.storage.get_object_version(path, delete_markers[0].version_id)
-
-        errors = self.storage.remove_object_with_versions(path)
-
-        self.test_case.assertEqual([], list(errors))
-        self.test_case.assertFalse(self.storage.exists(path))
-        self.test_case.assertEqual([], list(self.storage.list_object_versions(path)))
-        with self.test_case.assertRaises(FileNotFoundError):
-            self.storage.get_object_version(path, old_versions[0].version_id)
-
-    def test_remove_objects_for_missing_name_does_not_create_version_history(self) -> None:
-        path = self._track(PurePosixPath(f"dir{self.us}/missing-versioned.txt"))
-
-        errors = self.storage.remove_objects([path])
-
-        self.test_case.assertEqual([], list(errors))
-        self.test_case.assertEqual([], list(self.storage.list_object_versions(path)))
-
-    def test_invalid_names_raise_for_version_methods(self) -> None:
-        self.test_case.assertRaises(ValueError, self.storage.list_object_versions, "/")
-        self.test_case.assertRaises(ValueError, self.storage.get_object_version, "/", "v1")
-        self.test_case.assertRaises(ValueError, self.storage.get_object_version_stream, "/", "v1")
-        self.test_case.assertRaises(ValueError, self.storage.remove_object_with_versions, "/")
 
 
 class IBucketTester:  # pylint: disable=too-many-public-methods
